feat: Update icon path and fix Geek Uninstaller functionality
This commit updates the icon path in the main window's BrowserWindow options to use the correct path. It also fixes the functionality to launch Geek Uninstaller from the application. The `geekUninstaller.js` file is responsible for executing the `geek.exe` file. When the `launch-geek-uninstaller` button is clicked on the uninstaller page, the `ipcRenderer` triggers the `launch-geek-uninstaller` event, which executes Geek Uninstaller. Any encountered errors are logged to the console.